+++ src/io.cc
@@ -24,8 +24,8 @@ void dump(Data &data)
 #ifdef DUMP
   static size_t frame = 0;
 
-  const double t = (double)data;
-  const size_t n = (size_t)data;
+  const double t = data;
+  const size_t n = data;
   const size_t m = NVAR;
   const void  *h = data.host();
